+++ symbolic/cartpole_abstract.py
@@ -66,10 +66,10 @@ class CartPoleEnv_abstract(gym.Env):
         self.x_threshold = 0.05  # 2.4
 
         # Angle limit set to 2 * theta_threshold_radians so failing observation is still within bounds
-        high = np.array([self.x_threshold * 2, np.finfo(np.float32).max, self.theta_threshold_radians * 2, np.finfo(np.float32).max])
+        high = np.array([self.x_threshold * 2, np.finfo(np.float64).max, self.theta_threshold_radians * 2, np.finfo(np.float64).max])
 
         self.action_space = spaces.Discrete(2)
-        self.observation_space = spaces.Box(-high, high, dtype=np.float32)
+        self.observation_space = spaces.Box(-high, high, dtype=np.float64)
 
         self.seed()
         self.viewer = None
